+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - The app will now automatically import tracks from configurable folders
 - Fix y-range for elevation graphs
 - Display total ascent and total descent
-- Add average heart rate and total heart rate to statistics
+- Add average heart rate, total heart rate, average ascend and total ascend to statistics
 
 # 2024.12.1